+++ src/Transformer/ObjectToObjectMetadata/Implementation/ObjectToObjectMetadataFactory.php
@@ -214,15 +214,15 @@ final class ObjectToObjectMetadataFactory implements ObjectToObjectMetadataFacto
 
                 if ($targetWriteMode === WriteMode::None) {
                     continue;
-                } else {
-                    $targetWriteName = $targetWriteInfo->getName();
-                    $targetWriteVisibility = match ($targetWriteInfo->getVisibility()) {
-                        PropertyWriteInfo::VISIBILITY_PUBLIC => Visibility::Public,
-                        PropertyWriteInfo::VISIBILITY_PROTECTED => Visibility::Protected,
-                        PropertyWriteInfo::VISIBILITY_PRIVATE => Visibility::Private,
-                        default => Visibility::None,
-                    };
                 }
+                $targetWriteName = $targetWriteInfo->getName();
+                $targetWriteVisibility = match ($targetWriteInfo->getVisibility()) {
+                    PropertyWriteInfo::VISIBILITY_PUBLIC => Visibility::Public,
+                    PropertyWriteInfo::VISIBILITY_PROTECTED => Visibility::Protected,
+                    PropertyWriteInfo::VISIBILITY_PRIVATE => Visibility::Private,
+                    default => Visibility::None,
+                };
+
             }
 
             // get source property types